requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
python-escpos>=3.0
pywin32>=305
tkinterdnd2>=0.3.0
